refactor(asset-editor): name the pure-core module localAsset
Every export constructs, compares, or merges the asset being edited.
House rule: name files after their domain concept, not utils.
